+++ lib/knapsack_pro/regular_allocator_builder.rb
@@ -1,11 +1,10 @@
 # frozen_string_literal: true
 
 module KnapsackPro
-  class AllocatorBuilder < BaseAllocatorBuilder
+  class RegularAllocatorBuilder < BaseAllocatorBuilder
     def allocator
-      KnapsackPro::Allocator.new(
-        fast_and_slow_test_files_to_run: fast_and_slow_test_files_to_run,
-        fallback_mode_test_files: fallback_mode_test_files,
+      KnapsackPro::RegularAllocator.new(
+        test_suite: test_suite,
         ci_node_total: env.ci_node_total,
         ci_node_index: env.ci_node_index,
         repository_adapter: repository_adapter,