+++ spec/knapsack_pro/crypto/decryptor_spec.rb
@@ -15,7 +15,9 @@ describe KnapsackPro::Crypto::Decryptor do
   let(:decryptor) { described_class.new(test_files, encrypted_test_files) }
 
   describe '.call' do
-    subject { described_class.call(test_files, encrypted_test_files) }
+    let(:test_suite) { instance_double(KnapsackPro::TestSuite, calculate_test_files: instance_double(KnapsackPro::TestSuite::Result, test_files: test_files)) }
+
+    subject { described_class.call(test_suite, encrypted_test_files) }
 
     before do
       expect(KnapsackPro::Config::Env).to receive(:test_files_encrypted?).and_return(test_files_encrypted?)